+++ consensus/src/processes/dagknight/mod.rs
@@ -9,6 +9,7 @@ use crate::processes::ghostdag::ordering::SortableBlock;
 
 mod appendable_segment_tree_api;
 mod appendable_segment_tree_impl;
+mod segment_tree_ops;
 pub use appendable_segment_tree_api::{AppendableSegmentTreeApi, Bucket, bucket_for_score};
 pub use appendable_segment_tree_impl::AppendableSegmentTree;
 pub mod manager;
